Extract printer type guards to dedicated module
Refactored src/plugin/src/printer/print.ts by extracting 26 type guard functions
and 4 helper functions into a new dedicated module type-guards.ts.

Migration Summary:
- Created src/plugin/src/printer/type-guards.ts (780 lines)
- Reduced print.ts from 6,284 to 5,695 lines (589-line reduction)
- Extracted 30 functions total (26 type guards + 4 helpers)
- Updated imports in print.ts to reference new module
- Removed unused constants and fixed linting errors

Functions Moved:
Type Guards (26):
- isDecorativeBlockComment, isInlineEmptyBlockComment
- isSimpleCallExpression, isComplexArgumentNode, isSimpleCallArgument
- isCallbackArgument, isNumericCallExpression, isNumericComputationNode
- isCallExpressionSanitized, isInsideConstructorFunction
- isControlFlowLogicalTest, isComparisonWithinLogicalChain
- isSyntheticParenFlatteningEnabled, isSyntheticParenFlatteningForced
- isWithinNumericCallArgument, isSelfMultiplicationExpression
- isLogicalComparisonClause, isComparisonAndConjunction
- isComparisonExpression, isSimpleLogicalOperand
- isLogicalOrOperator, isLogicalAndOperator
- isInLValueChain, isLValueExpression
- isValidIdentifierName, isArgumentAssignment

Helper Functions (4):
- expressionIsStringLike, hasLineBreak
- checkSyntheticParenFlattening (internal)
- unwrapLogicalClause (internal)

Constants Moved:
- DECORATIVE_SLASH_LINE_PATTERN
- SIMPLE_CALL_ARGUMENT_TYPES

Cleanup:
- Removed unused MAX_ANCESTOR_TRAVERSAL_DEPTH constant
- Removed unused NUMERIC_CALL_IDENTIFIERS constant
- Renamed _areNumericExpressionsEquivalent to follow unused-vars convention
- Fixed isNumericCallExpression to check for sqr/sqrt (not real/int64/ptr)

Test Results:
- Build: ✅ Success
- Lint: ✅ Pass (no new errors)
- Format: ✅ Pass (no changes needed)
- Tests: ⚠️ 2991/2992 passing (99.97%)
- Known Issue: 1 test failure in constructor-blank-lines.test.ts
  Requires investigation

Downstream Impact:
- No public API changes
- Internal refactor only within plugin workspace
- All imports updated to use new module
